#270 (New Frontend) Generalize simple visualization plugins: - removed redundant code - adapted existing plugins to properly use generalized version
