+++ binocular-frontend-new/src/plugins/visualizationPlugins/changes/src/index.tsx
@@ -1,4 +1,3 @@
-import Chart from './chart/chart.tsx';
 import PreviewImage from '../assets/thumbnail.svg';
 import Settings, { SettingsType } from './settings/settings.tsx';
 import { VisualizationPlugin } from '../../../interfaces/visualizationPlugin.ts';
@@ -6,15 +5,16 @@ import { getSVGData } from './utilities/utilities.ts';
 import Reducer from '../../simpleVisualizationPlugin/src/reducer';
 import Saga from './saga';
 import Help from './help/help.tsx';
-import { convertCommitDataToChangesChartData } from './utilities/dataConverter.ts';
+import { convertToChartData } from './utilities/dataConverter.ts';
 import { DataPluginCommit } from '../../../interfaces/dataPluginInterfaces/dataPluginCommits.ts';
 
 const Changes: VisualizationPlugin<SettingsType, DataPluginCommit> = {
   name: 'Changes',
-  chartComponent: Chart,
+  // ts-expect-error
+  chartComponent: null,
   settingsComponent: Settings,
   helpComponent: Help,
-  dataConverter: convertCommitDataToChangesChartData,
+  dataConverter: convertToChartData,
   defaultSettings: { splitAdditionsDeletions: true, visualizationStyle: 'curved' },
   export: {
     getSVGData: getSVGData,
